[OM-62701]: Do not convert RI.numberOfCoupons to int now that the total number of coupons is double.

+++ com.vmturbo.cost.component/src/main/java/com/vmturbo/cost/component/reserved/instance/AbstractReservedInstanceStore.java
@@ -168,7 +168,7 @@ public abstract class AbstractReservedInstanceStore {
                         riBuilder.getReservedInstanceBoughtInfoBuilder()
                                 .getReservedInstanceBoughtCouponsBuilder()
                                 .setNumberOfCouponsUsed(coupons)
-                                .setNumberOfCoupons((int)coupons);
+                                .setNumberOfCoupons(coupons);
                     } else {
                         double coupons = riToUndiscoveredAccountUsage.getOrDefault(riBuilder.getId(),
                                 0d);
